feat: BaseFragment
